requirement.txt
aiofiles==22.1.0
argcomplete==1.10.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
chardet==3.0.4
compressed-rtf==1.0.6
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
httptools==0.4.0
IMAPClient==2.1.0
lxml==4.9.1
multidict==6.0.2
olefile==0.46
pdfminer.six==20191110
Pillow==9.2.0
pycryptodome==3.15.0
python-dotenv==0.21.0
python-pptx==0.6.21
pytz-deprecation-shim==0.1.0.post0
sanic==22.6.2
sanic-routing==22.3.0
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
textract==1.6.5
tzdata==2022.2
tzlocal==4.2
ujson==5.4.0
uvloop==0.16.0
websockets==10.3
xlrd==1.2.0
XlsxWriter==3.0.3